moved trialDaysRemaining and elapsedDays functions into their own files in the freemium folder
